add aidl_interface_headers module type
As part of the go/roboleaf conversion, all inputs to a rule need to be
explicitly provided to Bazel to satisfy sandboxing requirements. The
include_dirs property of aidl_interfaces passes directory paths to the
AIDL compiler to locate imports in AIDL files, but since Soong
sandboxing allows access of the entire source tree at rule execution
time, the sources used in the include_dirs property are not explicity
provided to the build system.

The aidl_interface_headers module type acts like a filegroup and wraps
a collection of AIDL sources and their import directory to track the
inputs for AIDL compilation actions. These inputs are then recorded as
implicits to the Soong rule as well. This explicit file information
will be used to allow Bazel to perform fully sandboxed actions.

The include_dirs property will be removed in a future commit in Internal
since there are some modules there that still use it.

Bug: 229251008
Test: m networkstack-aidl-interfaces
Change-Id: I8742645dca1b1054f4e3df63bfc5cfebb2c2b181