Convert versions_with_info and stability prop to Bazel
Stability prop is needed for aidl interfaces with ndk binding.

Versions_with_info is a new prop that will eventually replace versions and imports props. Since the way it works is very similar, this CL also refactor the existing implementation to look up import to be reusable in two scenarios.

Test: bp2build tests in Soong repo
Bug: 235113507
Change-Id: Ib0d0a87d559ae53b47a837901bddf29ea611c4a9