Add empty helper methods to enforce permissions
Previously, permissions were enforced within the onTransact method of
the Stub. Because the reference to a Stub can also be acquired locally
(that is, within the same process), it is possible to bypass these
checks. Introduce a new auxiliary method for each AIDL method which is
permission-annotated. This method will automatically implement the
permission checks declared in the annotation. It is expected (i.e.,
enforced via a lint) that the service implementation will call the
auxiliary method as its first statement. For example,

    @Override
    @EnforcePermission("READ_PHONE_STATE")
    public int myAPI() {
      myAPI_enforcePermission();
      // business logic
    }

If one argument of the method is an AttributionSource, it is considered
to be the source of the call. It is expected to be pass down as argument
to the auxiliary method, and will be used to authorize the action.

This change only creates empty helper methods. The services that already
rely on the annotations will first be updated to use these methods. A
secondary change will be used to move the permission checks from
onTransact into the new helper methods.

Test: atest aidl_integration_test
Bug: 241171714
Ignore-AOSP-First: Refactor internal first
Change-Id: I3a553d9fbb7193b5159885548086f7a2ef679c09