add frozen bool to aidl_interface
This will make it easier to understand the current state of an
aidl_interface module from the Android.bp file.
If `frozen: true`, then there are no changes between the last frozen
version and the next version.
If `frozen: false`, then there are changes to the next version of the
interface that has not been frozen.

Test: aidl_test
Bug: 231903487

Change-Id: I38d27bcdd47d57832a1a45c35e4c11b9a92ad77a