Ignore AttributionSource argument for @EnforcePermission
If a method is annotated with @EnforcePermission and one if its argument
is of type AttributionSource, it was asssumed that this argument would
be used for the verification of permissions. This case was introduced to
support IBluetooth.aidl which uses this convention.

Remove the support of this case as this is a temporary solution. A
long-term solution would likely involve the sharing of the
AttributionSource at a lower-level, similarly to WorkSource.

aidl-test-interface-permission-java-source is updated because of the new
implementation.

aidl-test-interface-platform-java-source is an outdated golden output.
It is removed.

Bug: 279811284
Test: atest aidl_integration_test
Test: atest EnforcePermissionTests
(cherry picked from https://android-review.googlesource.com/q/commit:c37da5ffe425a3efc8cb2adb685b599d5efe51fb)
Merged-In: I36d7b855ade20d205d984783f4ff43904f424e6f
Change-Id: I36d7b855ade20d205d984783f4ff43904f424e6f