Make aidl phony mod responsible for bp2build
This effective no-op refactoring prepares for allowlist v2. Allowlist v2
will require that, to be bp2build-on, a module's dependencies must be
bp2build-on. Previously, the phony module created by aidl_inteface was
not the same module that generated its bp2build definition, so any rdeps
of that phony module could not possibly be bp2build-on. This change
remedies this.

Fixes: 301676937
Test: m nothing
Test: Presubmits
Change-Id: Id2676856dbe967a629a0af10b62c3c0f62a65a06