Remove TestVndkRequiresFrozen test
TestVndkRequiresFrozen test was introduced to test frozen flag within
VNDK enabled aidl_interface, but this test was running on unexpected
environment, because device vndk version and platform vndk version was
not set. This makes test fail when build system is updated with VNDK
deprecation. As this test is no longer required with VNDK deprecation,
it would be better to remove test rather than fix the problem. This
change removes TestVndkRequiresFrozen from AIDL tests.

Bug: 316829758
Test: m nothing --no-skip-soong-tests passed
Change-Id: I79ab29949cdc6e4df4ac3c20e4749f74ea6ce95e