Add test for FixedSize parcelables with double and long fields
Since `double` and `long long` are aligned to 4 bytes on x86
(instead of 8 as expected) parcelables with these fields may
have different padding on different archs unless it's
handled explicitly. This commit adds a test case for this to
ensure that memory layouts match for the NDK and rust
backends. The NDK backend currently handles this by manually
aligning fields while the rust backend will manually insert
padding in a follow-up CL.

Bug: 294698705
Test: m aidl_integration_test
Change-Id: I67a0e90ba4c7b87adb1314b5e62c5274c5ef1682