Add bounds check to generated binder onTransact
Summary: Add a bounds check to generated binder code for out arrays. The maximum size of a binder transaction is only 1MB, so it should be impossible to ask for a result buffer so large. It's slightly difficult to estimate the true size of the allocation before doing it, but a maximum of 1 million elements is certainly larger than actually usable.

Test: atest aidl_unittests aidl_integration_test

Change-Id: Iac91c24c0a50938eeb31f3ec833c27e844cb4014
Signed-off-by: Abhishek Gadewar <abhishekgadewar@meta.com>