Update test for build system change
build/soong/rust is being changed to use
`--extern force:foo=path/to/foo.dylib` to decrease the number of patches
we need to carry on 3p crates by allowing `dylib`s to be built out of
`#![no_std]` code naturally. The flag check here needed to be changed to
accept that flag.

Bug: 168927099
Test: m blueprint_tests
Change-Id: Id271af6776f31dfed25c355e3ad7d7da9b2deb1d