Implement aidl update-api in soong_ui
We want to make the source tree read-only during the build, but
update-api writes to the source tree. Rewrite it so that it just builds
an aidl_update_api.txt file that contains information about all the
actions it wants to do, and then soong_ui will read that file and run
the scripts in it after the build is over and the read-only source tree
sandbox has been exited.

Bug: 200803952
Test: m androidx.core.uwb.backend.aidl_interface-(freeze|update)-api
Change-Id: I051f57aa97b96678ac6741584f0f463ca5fd54b1