Add support for annotated generic type parameters
This change extends the AIDL grammar to allow annotations on generic type parameters, for example <@FixedSize T>.

The AIDL compiler now parses these annotations and validates that any concrete type argument used to instantiate the generic type meets the specified constraints, failing with a compile time error if they do not.

This is specifically useful to MQDescriptor.aidl which requires the type
that's used with it to be @FixedSize, but without this new annotation
the user wouldn't find that out until they tried to use it with libfmq.

MQDescriptor can now be defined as:
parcelable MQDescriptor<@FixedSize T, Flavor> { ... }

Test: atest aidl_unittests
Bug: None
Flag: EXEMPT no users

Change-Id: I3acbe8bf8086fbee93367f463cd2cf205742610b