Reapply "Implement aidl update-api in soong_ui"
This reverts commit d85b417356f80018d73b25aa3852ab88b85852b9.

We want to make the source tree read-only during the build, but
update-api writes to the source tree. Rewrite it so that it just builds
an aidl_update_api.txt file that contains information about all the
actions it wants to do, and then soong_ui will read that file and run
the scripts in it after the build is over and the read-only source tree
sandbox has been exited.

Bug: 200803952
Test: m androidx.core.uwb.backend.aidl_interface-(freeze|update)-api
Test: m nothing --no-skip-soong-tests
Change-Id: I770c24cb819785a632d001bdc956f24875738511