Static assertions for tags of FixedSize unions in Rust
FixedSize types are required to use a specific layout that matches the
layout used by C++, and in the case of unions, this layout requires the
tag to be stored as a single byte at offset zero. To validate that,
insert static assertions that constructing each union variant produces a
value with the tag in the right place.

This adds a public `tag()` method to FixedSize unions. I made it public,
because maybe downstream crates want to use it. It seems useful enough.

Bug: 448853028
Test: presubmit
Change-Id: Ibbc2f0075df00db27a90a6f7c46c78ba587ae03d
Signed-off-by: Alice Ryhl <aliceryhl@google.com>