Validate the fixed sized array first
Signature() and ToString() expect to be able to evaluate fixed sized
array's dimensions to print the type. Those methods are often used for
error messages like when evaluating the number of type parameters in a
generic type. So we make sure we check those dimensions before checking
higher level details about the type.

Bug: 430478900
Test: aidl_parser_fuzzer && aidl_unittests
Flag: BUGFIX

Change-Id: I6996d449146aeeadd1b9041059d21cb11010b7d3